refactors blocks fetcher: repackage methods
